requirements.txt
Flask==3.0.0
pymongo==4.6.1
PyJWT==2.8.0
zeroconf==0.131.0
Werkzeug==3.0.1 
